requirements.txt
FastAPI
gunicorn
line-bot-sdk
langchain
